refactor(insights): co-locate trends chart transforms with TrendsLineChart
Move trendsChartTransforms.{ts,test.ts} from lib/charts/transforms/ to
sit next to TrendsLineChart.tsx — the only consumer.
