+++ frontend/src/scenes/trends/viz/trends-line-chart/TrendsLineChart.tsx
@@ -2,7 +2,6 @@ import { useValues } from 'kea'
 import posthog from 'posthog-js'
 import { useCallback, useMemo, type ErrorInfo } from 'react'
 
-import { buildTrendsChartConfig, buildTrendsSeries } from 'lib/charts/transforms/trendsChartTransforms'
 import { createXAxisTickCallback } from 'lib/charts/utils/dates'
 import { buildTheme } from 'lib/charts/utils/theme'
 import { DEFAULT_Y_AXIS_ID, LineChart, ReferenceLines, ValueLabels } from 'lib/hog-charts'
@@ -26,6 +25,7 @@ import { goalLinesToReferenceLines } from './goalLinesAdapter'
 import { handleTrendsLineChartClick } from './handleTrendsLineChartClick'
 import { TrendsAlertOverlays } from './TrendsAlertOverlays'
 import { buildTrendsYTickFormatter } from './trendsAxisFormat'
+import { buildTrendsChartConfig, buildTrendsSeries } from './trendsChartTransforms'
 import type { TrendsSeriesMeta } from './trendsSeriesMeta'
 import { TrendsTooltip } from './TrendsTooltip'
